requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
python-utils==2.3.0
requests==2.22.0
Werkzeug==1.0.0
pyOpenSSL==19.0.0
Flask-Login==0.5.0
oauthlib==3.1.0
python-dotenv==0.14.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
email-validator==1.1.1
